requirements.txt
robin_stocks
mysql-connector-python
pandas
scipy
h5py==2.10.0
keras==2.1.6
tensorflow==1.5.0
matplotlib
sklearn
ta
